[Tests] Re-enable the tests for integer overflow diagnostics
This commit splits up the test suite 'diagnostic_constant_propagation_propagation_int.swift'
into two files one each for 64bit and 32bit architectures.
The tests check the diagnostics for arithmetic and bitwise operations on
'Int' and 'UInt' values. The size of these numeric types depend on the word length of
the target architecture. Most of the tests added in this commit were present
earlier but were xfailed.